chore: move files out of base_keyboard_nav
