+++ src/flyout_cursor.js
