+++ src/gesture_monkey_patch.js
