+++ src/line_cursor.js
